+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityApplicationModel.java
@@ -22,6 +22,7 @@ import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.model.DefaultEntityApplicationModel;
 
 import java.util.Collection;
+import java.util.prefs.Preferences;
 
 /**
  * A Swing implementation of {@link is.codion.framework.model.EntityApplicationModel}
@@ -36,5 +37,16 @@ public class SwingEntityApplicationModel
 	 */
 	public SwingEntityApplicationModel(EntityConnectionProvider connectionProvider, Collection<SwingEntityModel> entityModels) {
 		super(connectionProvider, entityModels);
+	}
+
+	/**
+	 * Instantiates a new {@link SwingEntityApplicationModel}
+	 * @param connectionProvider the connection provider
+	 * @param entityModels the entity models
+	 * @param preferences the {@link Preferences} instance to use
+	 */
+	public SwingEntityApplicationModel(EntityConnectionProvider connectionProvider, Collection<SwingEntityModel> entityModels,
+																		 Preferences preferences) {
+		super(connectionProvider, entityModels, preferences);
 	}
 }