+++ changelog.md
@@ -13,6 +13,7 @@ Codion Change Log
 - JsonPreferencesStore bug fixed, lock file now cleanup up on save.
 ### is.codion.framework.model
 - EntityApplicationModel.version() removed.
+- DefaultEntityApplicationModel, Preferences constructor parameter added.
 ### is.codion.swing.framework.ui
 - EntityTablePanelPreferences bug fixed, column width preferences got lost for panels that were never initialized, now sets the column width as well as preferred width when applying preferences.
 - EntityTablePanelPreferences, auto-resize-mode preference added.